Add additional error handling
When receiving an attachment, check to see whether there are no images or more than 1 image. If these conditions are met, an error will be thrown. Otherwise, bot will reply with a temporary confirmation message to allow users to know if their attachment is accepted. Once text is extracted, the bot will edit the same confirmation message with the result, preventing message cluttering.